Move openstack nav dest to instance.openstack
There were several navigation destinations that were only available for
openstack type instances. Instead of checking for correct class type in
the navigation destination, move the classes and register them against
OpenStackInstance only.

FIXES RHCFQE-3842
